Small rewrite of the  reactor interface.
Recently, I had introduced addReader/addWriter to make the reactor 
interface closer to twisted's reactor interface, but we *have* 
to look inside tSMEventId for the mode on Unix; the attempt to guess 
whether POLLIN or POLLOUT couldn't be salvaged - the code was getting 
too ugly.

Besides, not all code had been converted to use addReader/addWriter 
anyway. 

So we're back to the old ways.


git-svn-id: file:///Library/svn/aculab/trunk@359 cae33c24-edb7-11db-8150-7705e9bc5c8f