Another rewrite of the reactor interface.
Basically, the old idea to always pass around tSMEventIds doesn't work, because the *!&%$! VMPs etc. manage their own events, and when the the stopped event is processed, the event information in the VMPs is bogus, and we cannot remove the fds from the reactor. The only way around this is to store the fds least for the VMPs.

The speech channels manage their own events and can mostly stay as is.


git-svn-id: file:///Library/svn/aculab/trunk@373 cae33c24-edb7-11db-8150-7705e9bc5c8f