Unify CallReactor and SpeechReactor - on v6, we only need a single reactor and that is indeed much better (no locking required).
Removed CallReactor, renamed SpeechReactor to Reactor.

For the old drivers, we start a thread that sends the call events to the real reactor. 

The combination v5 and Windows is currently unsupported. It's not difficult to repair thhat, but I can't test it here.

This change introduces one known bug: with v5 and the background thread, unused pipes will not be cleaned up. This is only a practical problem when one uses multiple reactors, but a bug nonetheless.


git-svn-id: file:///Library/svn/aculab/trunk@377 cae33c24-edb7-11db-8150-7705e9bc5c8f