Boy oh boy.
The unification of the event loops was way trickier than I thought it 
would be: there is a chicken-and-egg problem with call event 
dispatching: 

Basically, the call knows its handle only after openin/openout, but the 
global dispatcher may pickup events for the handle before the call was 
added to the CallEventThread. This must have always been a problem, but 
it was probably only visible at high load and it went unnoticed.


git-svn-id: file:///Library/svn/aculab/trunk@378 cae33c24-edb7-11db-8150-7705e9bc5c8f